fix: use correct status code for throwing media error
